pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Darf Vermieter Das? Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "cryptography>=44.0.1",
    "django>=5.1.6",
    "google-genai>=1.2.0",
    "openai>=1.61.0",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pyjwt>=2.10.1",
    "pyotp>=2.9.0",
    "python-dotenv>=1.0.1",
    "python-magic>=0.4.27",
]
